+++ CHANGELOG.md
@@ -5,11 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [v2.1.0] - 2021-06-16
 ### Added
 - New `logStderr` option for `shell` methods, intended to help debugging by printing out the stderr of a successful
 `shell` method execution - default value `false`.
 ### Changed
 - Error returned when parsing invalid JSON for an evaluation improved to help debugging.
+
 ## [v2.0.0] - 2021-04-15
 ### Changed
 - **BREAKING CHANGE** Project's Go code restructured, limited exposed packages. See [the migration
@@ -206,7 +209,9 @@ command and `evaluate` commands; default `3000` milliseconds.
 - Deploy image to Docker Hub.
 
 [Unreleased]:
-https://github.com/jthomperoo/custom-pod-autoscaler/compare/v2.0.0...HEAD
+https://github.com/jthomperoo/custom-pod-autoscaler/compare/v2.1.0...HEAD
+[v2.1.0]:
+https://github.com/jthomperoo/custom-pod-autoscaler/compare/v2.0.0...v2.1.0
 [v2.0.0]:
 https://github.com/jthomperoo/custom-pod-autoscaler/compare/v1.1.0...v2.0.0
 [v1.1.0]: